reorder timdex functions and organize under extras
